backend/requirements.txt
aiosqlite==0.19.0
antiorm==1.2.1
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.1
db==0.1.1
dj-rest-auth==5.0.1
Django==4.2.6
django-cors-headers==4.3.0
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
idna==3.4
oauthlib==3.2.2
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0
python-http-client==3.3.7
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
